# digiguide/docs/api_spec.yaml
openapi: 3.0.3
info:
  title: DigiGuide API
  description: 数码指南APP后端API接口文档 - SN序列号查询和电池健康度分析
  version: 3.1.0
  contact:
    name: DigiGuide Team
    email: support@digiguide.com

servers:
  - url: https://api.digiguide.com/v1
    description: 生产环境
  - url: http://localhost:8080/v1
    description: 开发环境

tags:
  - name: SN Query
    description: SN序列号查询相关接口
  - name: Battery Analysis
    description: 电池健康度分析相关接口
  - name: Brand
    description: 品牌信息相关接口

paths:
  /sn/decode:
    get:
      tags:
        - SN Query
      summary: 解码SN序列号
      description: 自动识别品牌并解码SN序列号，返回生产日期信息
      operationId: decodeSN
      parameters:
        - name: sn
          in: query
          required: true
          description: 序列号
          schema:
            type: string
            example: C3LXK2XXXXX
        - name: brand
          in: query
          required: false
          description: 指定品牌（可选）
          schema:
            $ref: '#/components/schemas/Brand'
      responses:
        '200':
          description: 解码成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SNDecodeResult'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /sn/validate:
    get:
      tags:
        - SN Query
      summary: 验证SN格式
      description: 验证SN序列号格式是否符合指定品牌规范
      operationId: validateSNFormat
      parameters:
        - name: sn
          in: query
          required: true
          schema:
            type: string
        - name: brand
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Brand'
      responses:
        '200':
          description: 验证结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean

  /sn/format/{brand}:
    get:
      tags:
        - SN Query
      summary: 获取品牌SN格式说明
      description: 获取指定品牌的SN序列号格式说明
      operationId: getFormatHint
      parameters:
        - name: brand
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Brand'
      responses:
        '200':
          description: 格式说明
          content:
            application/json:
              schema:
                type: object
                properties:
                  hint:
                    type: string

  /sn/warranty:
    get:
      tags:
        - SN Query
      summary: 获取保修信息
      description: 根据SN序列号估算保修状态
      operationId: getWarrantyInfo
      parameters:
        - name: sn
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 保修信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarrantyInfo'

  /sn/batch:
    post:
      tags:
        - SN Query
      summary: 批量解码SN
      description: 批量解码多个SN序列号
      operationId: batchDecodeSN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                snList:
                  type: array
                  items:
                    type: string
                  example: ["C3LXK2XXXXX", "R5CR70H1N4"]
      responses:
        '200':
          description: 批量解码结果
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SNDecodeResult'

  /battery/analyze:
    post:
      tags:
        - Battery Analysis
      summary: 上传bugreport文件分析
      description: 上传bugreport文件进行电池健康度分析
      operationId: analyzeBugreport
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 分析结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatteryAnalysisResponse'
        '400':
          description: 文件上传失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /battery/analyze/text:
    post:
      tags:
        - Battery Analysis
      summary: 分析bugreport文本
      description: 直接提交bugreport文本内容进行电池健康度分析
      operationId: analyzeBugreportText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
              description: bugreport文本内容
      responses:
        '200':
          description: 分析结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatteryAnalysisResponse'

  /battery/health:
    post:
      tags:
        - Battery Analysis
      summary: 计算电池健康度
      description: 根据电池原始数据计算健康度
      operationId: calculateHealth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatteryRawData'
      responses:
        '200':
          description: 健康度结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatteryHealthResult'

  /battery/reports:
    get:
      tags:
        - Battery Analysis
      summary: 获取报告列表
      description: 获取电池分析报告历史列表
      operationId: getReports
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 报告列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BatteryReportResponse'

  /battery/reports/{id}:
    get:
      tags:
        - Battery Analysis
      summary: 获取报告详情
      description: 获取指定电池分析报告详情
      operationId: getReportDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 报告详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatteryReportResponse'
        '404':
          description: 报告不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Battery Analysis
      summary: 删除报告
      description: 删除指定电池分析报告
      operationId: deleteReport
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  /brand/list:
    get:
      tags:
        - Brand
      summary: 获取品牌列表
      description: 获取支持的品牌列表
      operationId: getBrandList
      responses:
        '200':
          description: 品牌列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  brands:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          $ref: '#/components/schemas/Brand'
                        chinese:
                          type: string

  /brand/{brand}/info:
    get:
      tags:
        - Brand
      summary: 获取品牌信息
      description: 获取指定品牌的详细信息
      operationId: getBrandInfo
      parameters:
        - name: brand
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Brand'
      responses:
        '200':
          description: 品牌信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    $ref: '#/components/schemas/Brand'
                  chinese:
                    type: string
                  formatHint:
                    type: string
                  warrantyMonths:
                    type: integer

components:
  schemas:
    Brand:
      type: string
      enum:
        - XIAOMI
        - HUAWEI
        - OPPO
        - VIVO
        - APPLE
        - HONOR
        - SAMSUNG
        - LENOVO
        - HP
        - ASUS
        - DELL
        - APPLE_MAC
        - UNKNOWN
      description: 品牌枚举

    SNDecodeStatus:
      type: string
      enum:
        - SUCCESS
        - PARTIAL
        - FAILED
      description: 解码状态

    SNDecodeResult:
      type: object
      properties:
        brand:
          $ref: '#/components/schemas/Brand'
        rawSn:
          type: string
          description: 原始序列号
        factoryYear:
          type: integer
          description: 生产年份
        factoryMonth:
          type: integer
          description: 生产月份
        factoryWeek:
          type: integer
          description: 生产周次
        halfYear:
          type: string
          description: 半年（Apple）
        status:
          $ref: '#/components/schemas/SNDecodeStatus'
        errorMessage:
          type: string
          description: 错误信息
        productionDate:
          type: string
          description: 生产日期估算
        warrantyStatus:
          type: string
          description: 保修状态

    WarrantyInfo:
      type: object
      properties:
        sn:
          type: string
        brand:
          $ref: '#/components/schemas/Brand'
        purchaseDate:
          type: string
          format: date
        warrantyStartDate:
          type: string
          format: date
        warrantyEndDate:
          type: string
          format: date
        warrantyStatus:
          type: string
        remainingDays:
          type: integer

    BatteryRawData:
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        sn:
          type: string
        designCapacityMah:
          type: integer
          description: 设计容量(mAh)
        currentCapacityMah:
          type: integer
          description: 当前容量(mAh)
        chargeCounterMah:
          type: integer
          description: 当前电量计数器(mAh)
        cycleCount:
          type: integer
          description: 循环次数
        manufacturingDate:
          type: string
          description: 制造日期
        temperatureCelsius:
          type: number
          description: 温度(°C)
        screenOnTimeHours:
          type: integer
          description: 亮屏时间(小时)
        chargeCount:
          type: integer
          description: 充电次数

    ConfidenceLevel:
      type: string
      enum:
        - HIGH
        - MEDIUM
        - LOW
        - NONE
      description: 置信度级别

    HealthFactors:
      type: object
      properties:
        capacityRetention:
          type: number
          description: 容量保持率(0-1)
        cycleDecay:
          type: number
          description: 循环衰减(0-1)
        resistanceGrowth:
          type: number
          description: 内阻增长(0-1)
        thermalAging:
          type: number
          description: 温度老化(0-1)
        chargingDamage:
          type: number
          description: 充电损伤(0-1)
        availableFactors:
          type: integer
          description: 可用因子数量

    BatteryHealthResult:
      type: object
      properties:
        healthPercentage:
          type: number
          description: 综合健康度(0-100)
        grade:
          type: string
          enum: [A+, A, B, C, D, F]
          description: 健康度等级
        factors:
          $ref: '#/components/schemas/HealthFactors'
        diagnosisText:
          type: string
          description: 诊断文字
        suggestions:
          type: array
          items:
            type: string
          description: 使用建议
        estimatedResistanceMohm:
          type: number
          description: 估算内阻(mΩ)
        remainingLifespanMonths:
          type: integer
          description: 预估剩余寿命(月)
        confidence:
          $ref: '#/components/schemas/ConfidenceLevel'

    BatteryAnalysisResponse:
      type: object
      properties:
        rawData:
          $ref: '#/components/schemas/BatteryRawData'
        healthResult:
          $ref: '#/components/schemas/BatteryHealthResult'
        reportId:
          type: integer
          description: 报告ID
        analysisTime:
          type: string
          format: date-time

    BatteryReportResponse:
      type: object
      properties:
        id:
          type: integer
        brand:
          type: string
        model:
          type: string
        sn:
          type: string
        designCapacityMah:
          type: integer
        currentCapacityMah:
          type: integer
        cycleCount:
          type: integer
        healthPercentage:
          type: number
        grade:
          type: string
        reportTime:
          type: string
          format: date-time
        diagnosisText:
          type: string

    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误信息